Exception handling, common list response for sport-API, adding proxy API calls for https://api.statscore.com/v2/events/{event_id}/sub-participants https://api.statscore.com/v2/participants/{participant_id}/squad https://api.statscore.com/v2/participants https://api.statscore.com/v2/participants/{participant_id} https://api.statscore.com/v2/events/{event_id} https://api.statscore.com/v2/events

+++ src/main/java/io/limeup/flexbets/sport/api/SubParticipantController.java
@@ -35,7 +35,7 @@ public class SubParticipantController {
 
     @GetMapping("/{sub-participant_id}")
     public ResponseEntity<SubParticipantDTO> getSubParticipantById(
-            @PathVariable("sub-participant_id") Long subParticipantId) {
+            @PathVariable("sub-participant_id") Integer subParticipantId) {
         return ResponseEntity.ok(subParticipantService.getSubParticipantById(subParticipantId));
     }
 }
